240521. partition shuffle

+++ llama_train.py
@@ -313,7 +313,7 @@ def llama_finetune(
                 if args.combined:
                     data['predicted_know'] = data['predicted_know'][0:0 + 5] + data['predicted_know'][10:10 + 5]
                     if args.partition:
-                        n_partition_negative = args.n_hard_negative / 2
+                        n_partition_negative = int(args.n_hard_negative / 2)
                         top1_hard_negative_candidates = [item for item in data['predicted_know'][0:0+5] if item != data['gpt_selection']]
                         top1_hard_negative_candidates = top1_hard_negative_candidates[:n_partition_negative]
                         top2_hard_negative_candidates = [item for item in data['predicted_know'][10:10+5] if item != data['gpt_selection']]
@@ -330,7 +330,7 @@ def llama_finetune(
             else:
                 n_sampled_negative = args.n_sampled_negative
             if args.partition:
-                n_partition_sampled_negative = n_sampled_negative / 2
+                n_partition_sampled_negative = int(n_sampled_negative / 2)
                 tmp_know_1 = []
                 tmp_know_2 = []
                 while len(tmp_know_2) < n_partition_sampled_negative:
